Contract/plugin reserve

+++ libs/contract-oracle/contracts/AggregatorV3Interface.sol
@@ -57,8 +57,8 @@ interface AggregatorV3Interface {
 
   // Custom TC function.  Only the Oracle can convert from
   // timestamp to roundId, so we might as well encapsulate it here.
-  function getRoundFromTimestamp(int timestamp)
+  function getRoundFromTimestamp(uint timestamp)
     external
     view
-    returns (int256 answer);
+    returns (uint answer);
 }
